+++ ChangeLog.md
@@ -5,6 +5,7 @@
 - Adding consistent album artist support
 - Build support for libnpupnp 6.1.2
 - Build: Allow build of ffmpegthumbnailer without ffmpeg
+- Build: rework dependency handling
 - CI: Drop excluded ubuntu versions as CMake PPA now supports those combinations
 - CI: Pass publish-deb secrets via env instead
 - CI: Upload debs to pkg.gerbera.io